Changing how an offspring is evaluated
The offspring test should not always be added to the offspring population. In fact, when the test ends with a timeout (aka the test is too expensive) or has some internal exception should not be added to the population.

In my trial experiments, the class com.pmdesigns.jvc.tools.JVCParserTokenManager often leads to generating expensive tests (offspring) that should not be added to the population. With this commit, this minor issue should be fixed